Create separate test files for bindings value and field actions
